containers/streamlit/requirements.txt
streamlit>=1.17.0,<1.18
transformers[torch]>=4.26.0,<4.27
pillow>=9.4.0,<9.5
